:truck: :: [#90] FeatureLayer / BaseViewController와 BaseStoredViewController 분리
