+++ Projects/Feature/BaseFeature/Sources/Base/BaseStoredViewController.swift
@@ -1,0 +1,34 @@
+import UIKit
+
+open class BaseStoredViewController<Store: BaseStore>:
+    BaseViewController,
+    StoredViewControllable,
+    StoreBindable {
+
+    // MARK: - Properties
+
+    public let store: Store
+
+    // MARK: - Init
+
+    public init(store: Store) {
+        self.store = store
+        super.init(nibName: nil, bundle: nil)
+    }
+
+    required public init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    // MARK: - LifeCycle
+
+    open override func viewDidLoad() {
+        bindAction()
+        bindState()
+        super.viewDidLoad()
+    }
+
+    open func bindState() {}
+
+    open func bindAction() {}
+}
